Sql ast uses scalaz Equal.
Break Sql ast into multiple files.
