+++ frontend/src/main/scala/quasar/sql/JoinType.scala
@@ -1,0 +1,33 @@
+/*
+ * Copyright 2014–2016 SlamData Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package quasar.sql
+
+import quasar.Predef._
+
+import scalaz._
+
+sealed abstract class JoinType(val sql: String) extends Product with Serializable
+
+final case object LeftJoin extends JoinType("left join")
+final case object RightJoin extends JoinType("right join")
+final case object InnerJoin extends JoinType("inner join")
+final case object FullJoin extends JoinType("full join")
+
+object JoinType {
+  implicit val equal: Equal[JoinType] = Equal.equalRef
+  implicit val show: Show[JoinType] = Show.showFromToString
+}
